requirements.txt
Django==5.1
djangorestframework==3.12.4
django-environ==0.7.0
django-filter==21.1
django-autoslug==1.9.8
django-countries==7.2.1
flake8==3.9.2
Pillow==8.3.2
django-phonenumber-field==5.2.0
phonenumbers==8.12.33
psycopg2-binary==2.9.1
black==21.6b0
isort==5.9.3
djoser==2.2.3
PyJWT==2.9.0
djangorestframework-simplejwt==5.3.1
pytest-django==4.4.0
pytest-factoryboy==2.1.0
Faker==8.9.1
pytest-cov==2.12.1